Only reboot once...

+++ raspbernetes/bootstrap.d/04-install_prerequisites.sh
@@ -1,18 +1,24 @@
 #!/bin/bash
 set -euo pipefail
 
-echo "Updating core packages..."
-DEBIAN_FRONTEND=noninteractive apt-get update
+FILE="${KUBE_NODE_USER_HOME}/reboot"
+if [ ! -f "$FILE" ]; then
+    echo "Updating core packages..."
+    DEBIAN_FRONTEND=noninteractive apt-get update
 
-echo "Installing base packages..."
-DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends \
-apt-transport-https \
-software-properties-common \
-zip \
-jq \
-git \
-vim \
-curl
+    echo "Installing base packages..."
+    DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends \
+    apt-transport-https \
+    software-properties-common \
+    zip \
+    jq \
+    git \
+    vim \
+    curl
 
-DEBIAN_FRONTEND=noninteractive apt-get -y upgrade
-reboot
+    DEBIAN_FRONTEND=noninteractive apt-get -y upgrade
+    touch "${FILE}"
+
+    echo "Reboot in progress..."
+    reboot
+fi
